add(Repository): add a theaters repository
