+++ src/controllers/theater.controller.js
@@ -1,151 +1,32 @@
 const { checkValidation } = require('../middleware/validation.middleware');
-const { structureResponse } = require('../utils/common.utils');
-const { dbTransaction } = require('../db/db-connection');
 
-const TheaterModel = require('../models/theater.model');
-const TheaterSeatModel = require('../models/theaterSeat.model');
-const SeatType = require('../utils/enums/seatTypes.utils');
-const {
-    NotFoundException,
-    CreateFailedException,
-    UpdateFailedException,
-    UnexpectedException
-} = require('../utils/exceptions/database.exception');
+const TheaterRepository = require('../repositories/theater.repository');
 
 class TheaterController {
     getAllTheaters = async (req, res, next) => {
-        let theaterDuplicates = await TheaterModel.findAll();
-        if (!theaterDuplicates.length) {
-            throw new NotFoundException('Theaters not found');
-        }
-
-        let theaterList = {};
-
-        for (const theater of theaterDuplicates) {
-            const {seat_row, seat_number, seat_type, ...theaterDetails} = theater;
-            const theater_id = theater.theater_id;
-            if (!theaterList[theater_id]) {
-                theaterList[theater_id] = theaterDetails;
-                theaterList[theater_id].missing = [];
-                theaterList[theater_id].blocked = [];
-            }
-            if (seat_type === SeatType.Missing){
-                theaterList[theater_id].missing.push({ seat_row, seat_number });
-            } else {
-                theaterList[theater_id].blocked.push({ seat_row, seat_number });
-            }
-        }
-
-        theaterList = Object.values(theaterList);
-
-        const response = structureResponse(theaterList, 1, "Success");
+        const response = await TheaterRepository.findAll();
         res.send(response);
     };
 
     getTheaterById = async (req, res, next) => {
-        const theaterDuplicates = await TheaterModel.findOne({ theater_id: req.params.id });
-        if (!theaterDuplicates.length) {
-            throw new NotFoundException('Theater not found');
-        }
-
-        let theaterBody = {};
-
-        for (const theater of theaterDuplicates) {
-            const {seat_row, seat_number, seat_type, ...theaterDetails} = theater;
-            if (Object.keys(theaterBody).length === 0) {
-                theaterBody = theaterDetails;
-                theaterBody.missing = [];
-                theaterBody.blocked = [];
-            }
-            if (seat_type === SeatType.Missing){
-                theaterBody.missing.push({ seat_row, seat_number });
-            } else {
-                theaterBody.blocked.push({ seat_row, seat_number });
-            }
-        };
-
-        const response = structureResponse(theaterBody, 1, "Success");
+        const response = await TheaterRepository.findOne({ theater_id: req.params.id });
         res.send(response);
     };
 
     createTheater = async (req, res, next) => {
         checkValidation(req);
-
-        const {missing, blocked, ...theaterBody} = req.body;
-
-        await dbTransaction.beginTransaction();
-
-        const result = await TheaterModel.create(theaterBody);
-
-        if (!result) {
-            await dbTransaction.rollback();
-            throw new CreateFailedException('Theater failed to be created');
-        }
-
-        if (missing) {
-            for (let seat of missing) {
-                seat = seat.split("-"); // "A-11" -> ["A",11];
-                const theaterSeat = {
-                    theater_id: result.theater_id,
-                    seat_type: SeatType.Missing,
-                    seat_row: seat[0],
-                    seat_number: seat[1]
-                };
-                const success = await TheaterSeatModel.create(theaterSeat);
-                if (!success) {
-                    await dbTransaction.rollback();
-                    throw new CreateFailedException('Theater missing seat failed to be created');
-                }
-            }
-        }
-        if (blocked){
-            for (let seat of blocked) {
-                seat = seat.split("-"); // "A-11" -> ["A",11];
-                const theaterSeat = {
-                    theater_id: result.theater_id,
-                    seat_type: SeatType.Blocked,
-                    seat_row: seat[0],
-                    seat_number: seat[1]
-                };
-                const success = await TheaterSeatModel.create(theaterSeat);
-                if (!success) {
-                    await dbTransaction.rollback();
-                    throw new CreateFailedException('Theater blocked seat failed to be created');
-                }
-            }
-        }
-
-        await dbTransaction.commit();
-
-        const response = structureResponse(result, 1, 'Theater was created!');
+        const response = await TheaterRepository.create(req.body);
         res.status(201).send(response);
     };
 
     updateTheater = async (req, res, next) => {
         checkValidation(req);
-
-        const result = await TheaterModel.update(req.body, req.params.id);
-
-        if (!result) {
-            throw new UnexpectedException('Something went wrong');
-        }
-
-        const { affectedRows, changedRows, info } = result;
-
-        if (!affectedRows) throw new NotFoundException('Theater not found');
-        else if (affectedRows && !changedRows) throw new UpdateFailedException('Theater update failed');
-        
-        const response = structureResponse(info, 1, 'Theater updated successfully');
+        const response = await TheaterRepository.update(req.body, req.params.id);
         res.send(response);
     };
 
     deleteTheater = async (req, res, next) => {
-        const result = await TheaterModel.delete(req.params.id);
-        if (!result) {
-            throw new NotFoundException('Theater not found');
-        }
-
-        const response = structureResponse({}, 1, 'Theater has been deleted');
+        const response = await TheaterRepository.delete(req.params.id);
         res.send(response);
     };
 }
